Update Video Parser and general build fixes
Add support for "Video Session Parameters".
Updated .gitignore
Check if the provided file to be decoded is valid.
Build issues fixes.
Update the build documentation.
Remove the dependency on VK_NV_video_queue
Ensure the right vulkan driver version is in use
Updated the README and BUILD files for the release