Add support for auto format and size video update
Add support for video stream auto-resize and format update
Change the video format dynamically in the content display processor
Auto-scale the video image when the image height is bigger than the coded extent.
Determine if the video session needs to be recreated and create a new one if needed.
Determine if the video images need to be recreated on format or extent change.
Update images if the size is smaller or the format is incompatible with the content change update.