VLfix h264 pEncodeInfo->pSetupReferenceSlot->pNext chain includes unexpected VkStructureType
Details:
- [ VUID-VkVideoReferenceSlotInfoKHR-pNext-pNext ] vkCmdEncodeVideoKHR(): pEncodeInfo->pSetupReferenceSlot->pNext
chain includes a structure with unexpected VkStructureType VK_STRUCTURE_TYPE_APPLICATION_INFO. This error is based
on the Valid Usage documentation for version 290 of the Vulkan header. It is possible that you are using a struct
from a private extension or an extension that was added to a later version of the Vulkan header, in which case the
use of pEncodeInfo->pSetupReferenceSlot->pNext is undefined and may not work correctly with validation enabled.
The Vulkan spec states: Each pNext member of any structure (including this one) in the pNext chain must be either
NULL or a pointer to a valid instance of VkVideoDecodeAV1DpbSlotInfoKHR, VkVideoDecodeH264DpbSlotInfoKHR,
VkVideoDecodeH265DpbSlotInfoKHR, VkVideoEncodeH264DpbSlotInfoKHR, or VkVideoEncodeH265DpbSlotInfoKHR.