VLfix redundant semaphore
Redundant semaphore was signaled and has not since been waited on

Details:
- [ VUID-vkQueueSubmit-pCommandBuffers-00065 ] VK_OBJECT_TYPE_SEMAPHORE; vkQueueSubmit(): pSubmits[0].pSignalSemaphores[0]
(VkSemaphore 0x..[]) is being signaled by VkQueue 0x56045cf78050[], but it was previously signaled by VkQueue 0x56045cf78050[]
and has not since been waited on. The Vulkan spec states: Any calls to vkCmdSetEvent, vkCmdResetEvent or vkCmdWaitEvents that have been
recorded into any of the command buffer elements of the pCommandBuffers member of any element of pSubmits, must not reference any VkEvent
that is referenced by any of those commands in a command buffer that has been submitted to another queue and is still in the pending state.