VLfix dpb slot management
Details:
- Fixed encoder corruption because of dpb slots in conjuction with the following errors:
- Fixed the following validation errors (relavant to the same problem with the dpb slot management):
1) Validation Error: [ VUID-vkCmdEncodeVideoKHR-dpbFrameUseCount-08221 ] vkCmdEncodeVideoKHR(): frame in DPB slot 3
is referred to multiple times across pEncodeInfo->pSetupReferenceSlot and the elements of pEncodeInfo->pReferenceSlots.
The Vulkan spec states: All elements of dpbFrameUseCount must be less than or equal to 1;
2) Validation Error: [ VUID-vkCmdEncodeVideoKHR-pPictureResource-08219 ] vkCmdEncodeVideoKHR(): the bound video picture
resource specified in pEncodeInfo->pReferenceSlots[2].pPictureResource is not currently associated with the DPB slot index
specifed in pEncodeInfo->pReferenceSlots[2].slotIndex (3). The Vulkan spec states: The pPictureResource member of each element
of pEncodeInfo->pReferenceSlots must match one of the bound reference picture resource associated with the DPB slot index
specified in the slotIndex member of that element.
3) Validation Error: [ VUID-vkCmdBeginVideoCodingKHR-pPictureResource-07265 ] vkCmdBeginVideoCodingKHR():  DPB slot index 3
of VkVideoSessionKHR 0x..[] is not currently associated with the specified video picture resource: VkImage 0x..[], layer 0,
offset (x = 0, y = 0), extent (width = 1920, height = 1080). The Vulkan spec states: Each video picture resource specified
by any non-NULL pPictureResource member specified in the elements of pBeginInfo->pReferenceSlots for which slotIndex is not
 negative must match one of the video picture resources currently associated with the DPB slot index of pBeginInfo->videoSession
 specified by slotIndex at the time the command is executed on the device.