encoder: Make callbacks non-static
This vector was declared as static, but the lambdas capture `this`.
This broke CTS, since it may construct many VkVideoEncoder objects.
After the first VkVideoEncoder object is destroyed, the subsequent calls
to ProcessOrderedFrames() do not reconstruct this vector, while the
lambdas still use `this` from the first time it was created.

Fixes #11.