encoder: fix VkVideoPictureResourceInfoKHR::baseArrayLayer
If the image view is an array, the baseArrayLayer points into the image
view's layers. But in that path, each image view is only one layer, so
the baseArrayLayer should be 0.

If the image view is not an array, the baseArrayLayer should be 0, since
the image view has 0 layers.

Fixes VUID-VkVideoPictureResourceInfoKHR-baseArrayLayer-07175.