HEVC encode: support multiple slices
Also use a separate slice segment header structure for each slice. This
will be useful when subsequently adding support for intra-refresh, where
the valid API usage rules require that the slice corresponding to the
intra-refresh region being currently refreshed should be marked as an
intra slice, and the slice type is signaled in the slice segment header.

Signed-off-by: Srinath Kumarapuram <skr@nvidia.com>