Use the SHADERC_SHARED_LIBRARY variable/target for shaderc linkage
There are multiple paths in the build scripts for either finding shaderc
on the filesystem or fetching the source and building it, before linking
it into the sample applications. These paths use separate CMake
variable/target names, which makes it confusing to understand
dependencies and makes refactoring / bug-fixing difficult.

This change attempts to slightly improve this situation by consistently
using the SHADERC_SHARED_LIBRARY variable/target wherever linkage to
shaderc is necessary. This means removing the use of explicit linker
flags for linking on Linux and letting CMake infer the required linker
flags/options from the specified variable/target.

SHADERC_SHARED_LIBRARY has been chosen instead of SHADERC_LIB_SHARED
because the latter is a hard-coded name ("shaderc_shared") in the build
script and does not correspond to the actual library name in all cases
(for example, the library installed by using the package manager on
Ubuntu 24.04 is named libshaderc.so.1).

The SHADERC_LIB_SHARED_DIR variable was only used along with explicit
linker flags, to specify the directory to search for shaderc. With the
removal of explicit linker flags, this variable is no longer used.

Signed-off-by: Srinath Kumarapuram <skr@nvidia.com>