encode: simplify VkVideoEncodeFrameInfo::GetType()
This method currently assumes that the first element in the pNext chain
of VkVideoEncodeInfoKHR will be a codec-specific structure for the
picture info (e.g. VkVideoEncodeH264PictureInfoKHR). This assumption is
problematic because:

- it complicates the pNext chaining logic for other places in the code
  which need to insert one or more structures into the pNext chain for
  VkVideoEncodeInfoKHR.

- making assumptions within the application about the order of
  structures in the pNext chain, although not prohibited by the Vulkan
  specification, may lead to similar assumptions elsewhere and
  complicate code writing and comprehension.

This method aids VkVideoEncoderH264::GetEncodeFrameInfoH264() and
similar methods in the other codec-specific classes in checking that the
pointer to the base class points in fact to the actual codec-specific
data structure needed, instead of relying on runtime type information
(RTTI). An alternative and simple way to achieve this, as implemented in
this change, is to add private member to VkVideoEncodeFrameInfo for
holding the codec operation. This member is populated by the
constructor, based on the constructors of derived classes.

Signed-off-by: Srinath Kumarapuram <skr@nvidia.com>