encode: improve reset of VkVideoEncodeInfoKHR.pNext chain
Prior to this commit, the code in VkVideoEncodeFrameInfo::Reset() was
written with the assumption that the first element in the pNext chain of
VkVideoEncodeInfoKHR would be a codec-specific structure describing the
picture info (e.g. VkVideoEncodeH264PictureInfoKHR). Correspondingly,
the overridden versions of ::Reset() in the derived classes would simply
do:

    pictureInfo.pNext = nullptr;

to clear the pNext chain.

Commit 16607feb9b86984b53e1747332c1ca386d7ecd2d regressed this by
changing the pNext chaining behaviour in the ChainNextVkStruct() helper
to add structures at the head of the pNext chain rather than at the
tail. This meant that the codec-specific picture info structure was not
guaranteed to be the first structure in the pNext chain of
VkVideoEncodeInfoKHR, so the above assignment was not sufficient to
clear the pNext chain.

This change updates VkVideoEncodeFrameInfo::Reset() to clear the pNext
of encodeInfo (i.e. VkVideoEncodeInfoKHR) and then have it point to
codec-specific picture info structure in the overridden ::Reset()
provided by the derived class. This change also restores the
functionality of quantization maps which was broken by commit
16607feb9b86984b53e1747332c1ca386d7ecd2d.

Signed-off-by: Srinath Kumarapuram <skr@nvidia.com>