encode: simplify pNext chaining for the inline query structure
Previously, any code adding a structure to the pNext chain of
VkVideoEncodeInfoKHR had to be careful to preserve a codec-specific
picture as the first element in this chain, or restore it to that state
as soon as possible, because code elsewhere assumed it or depended on it
for correct operation. With the previous two commits, such assumptions
have been eliminated, making pNext chaining for VkVideoEncodeInfoKHR
much easier.

Signed-off-by: Srinath Kumarapuram <skr@nvidia.com>