encode: check QP map file size against number of frames being encoded
The file processing code expects a QP map to be available for each frame
being encoded but does not check that this is actually the cause.
Specifying a QP map file with an insufficient number of QP maps can
cause a segmentation fault / access violation. This commit adds the
method EncoderQpMapFileHandler::GetFrameCount() to allow checking that
the QP map file has a sufficient number of QP maps.

Signed-off-by: Srinath Kumarapuram <skr@nvidia.com>