encode: update H.264 level limits based on the profile
The level determination code in EncoderConfigH264::DetermineLevel() uses
a multipler of 1200 for the MaxBR and MaxCPB entries in the level limits
table, taken from table A-1 of the H.264 specification. However, this
multipler is to be used only for the Baseline, Constrained Baseline,
Main and Extended profiles. Profiles greater than these use a different
(larger) multipler, as specified in table A-2 of the H.264
specification. Due to the use of a smaller multipler, this code selects
a greater level than is strictly necessary in some situations, based on
the input settings.

This commit updates this code to use a multipler dependent on the codec
profile.

Signed-off-by: Srinath Kumarapuram <skr@nvidia.com>