encode: add EncoderConfigH264 method for profile and level determination
The codec profile and level are currently set in the following places:
- level in EncoderConfigH264::InitDpbCount()
- profile in EncoderConfigH264::InitSpsPpsParameters()

The profile and level are fairly basic encoding parameters and it is
desirable to set these as early as possible. This commit adds a method
to the EncoderConfigH264 class which sets the profile and level at an
early stage in the encoder configuration. This also provides the
opportunity to use the codec profile in the construction of the video
profile which will be used in physical device queries (currently, this
video profile uses a default codec profile).

The profile is more important than the level, so it is essential to set
these values in this order, as done within this change. Also, certain
level limits are profile-specific, and this ordering makes it possible
for a subsequent change to use the profile in the level determination
code.

Signed-off-by: Srinath Kumarapuram <skr@nvidia.com>